Fix Key Error when editing CD ROM path
lxml only accepts strings as parameters when looking
up xml elements. The port for remote ISO was being
passed in as an int, and a KeyError was thrown.

This patch just casts the ints to strings for the
lookups.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>
(cherry picked from commit f4006fee2a2626f115b074cbc95a02db0899923f)

Conflicts:
	src/kimchi/model/vmstorages.py

- assigned port to a variable <rodrigo.trujillo@linux.vnet.ibm.com>