Bugfix UI: Change button text to indicate user network is generating
When network is generating, if the network is not up, we bring the
network up first, then we create the network. During the time we bring
the network up(less than 10s), we change the text as well as the layout
of the button to indicate this process is working.

Signed-off-by: Wen Wang <wenwang@linux.vnet.ibm.com>
(cherry picked from commit 191c2f4f337577482a1e7be86293acebf6c4c659)