Change function that verifies network interface status
The old way does just consider if the interface has connectivity, not
the interface status. The new file can determine if the interface is up
or down and if cable is connected or not.

Signed-off-by: Ramon Medeiros <ramonn@linux.vnet.ibm.com>
(cherry picked from commit c456d43459b09fb831c04c938c4b4c8f8c49e495)