Issue #405 - Fix cancel button in edit guest storage tab
When the user is in storage tab in edit guest window, he can clicks on
'replace' button in order to change the ISO path. However, if user types
anything in the path input box and wants to cancel the changes, turns
out that the 'cancel' button was not restoring previous value. This patch
fixed this problem.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>
(cherry picked from commit 2524fb449b54212dc0882072f86303096d49a069)