Update iSCSI volume XML when creating a VM from an iSCSI pool
When creating a VM using a iSCSI volume, the volume XML was like below:

<disk type='volume' device='disk'>
  <driver name='qemu' type='raw'/>
  <source pool='TEST-ISCSI1' volume='unit:0:0:2' mode='host'/>
  <target dev='vda' bus='virtio'/>
  <address type='pci' domain='0x0000' bus='0x00' slot='0x05' function='0x0'/>
</disk>

But to attach a new iSCSI volume, the XML is:

<disk type='block' device='disk'>
  <driver name='qemu' type='raw'/>
  <source dev='/dev/disk/by-id/wwn-0x60000000000000000e00000000010001'/>
  <target dev='vdb' bus='virtio'/>
  <address type='pci' domain='0x0000' bus='0x00' slot='0x07' function='0x0'/>
</disk>

We should keep the same XML structure in both cases to avoid problems
while listing the storage volumes and getting their information.
To solve that, this patch uses the latter XML also while creating a VM
from a iSCSI volume.

It eliminates the needs to have the authentication setting in the
volume, because that the _get_storage_auth() function was replaced by
_get_volume_path()

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>
(cherry picked from commit e6931d7245cced2227f05c28742c77253bda413c)

Conflicts:
	src/kimchi/model/vms.py