PCI Passthru: Adjust the passthrough filter rules
In previous commits, we don't allow to passthrough device with children,
and only passthrough the children. It proves it's inflexible and less
useful. The PCI class code white list also filter out too much types of
devices, and there is no way to cleanly differenciate devices suitable
to passthrough using class code.

In this patch, we allow Kimchi to passthrough a parent PCI device. The
front-end can use the existing "_passthrough_group_by" filter to list
the affected children devices and other PCI devices in the same group,
as well as the children of those devices. We also drop the class code
white list, and filter out only PCI bridge and video cards.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>