PCI Passthru: Properly Parse NPIV Capable HBA Card Information
If the FC HBA card does not support NPIV, the capability XML is like
  <capability type='scsi_host'>
    <host>2</host>
    <capability type='fc_host'>
      <wwnn>20000090fa530215</wwnn>
      <wwpn>10000090fa530215</wwpn>
      <fabric_wwn>0</fabric_wwn>
    </capability>
  </capability>

If it supports NPIV, it's like
  <capability type='scsi_host'>
    <host>2</host>
    <capability type='fc_host'>
      <wwnn>20000090fa530215</wwnn>
      <wwpn>10000090fa530215</wwpn>
      <fabric_wwn>0</fabric_wwn>
    </capability>
    <capability type='vport_ops'>
      <max_vports>895</max_vports>
      <vports>0</vports>
    </capability>
  </capability>

The difference is that one more sub-capability with type of 'vport_ops'
is appended to the capability. If there is only one sub-capability,
after parsing of kimchi.xmlutils.dictize, the result is just a
dictionary. If there are multiple sub-capabilities, the result is a list of
dictionaries. Previously the code processing the parsing result assumes
that there is only one sub-capability, and so as the consumer of the
hostdev module. It would fail the consumer if there are multiple
sub-capabilities. In this patch the post-processing checks if the
parsing result is a list, and combines the information in the list into a
single dictionary, so that the consumer of hostdev module still gets
just a single dictionary even if there are multiple sub-capabilities.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>