Pass through libvirt error if storage create fails
A user experienced an error while creating a VM using
logical storage. If a disk size greater than the available
capacity of a pool is requested, the pool.createXML fails,
but the user gets the generic kimchi "unexpected error" message.
This patch puts a try block around the createXML call, and
catches the libvirt error so it can be shown to the user.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>
(cherry picked from commit 4c4e582cb0019900cf826ef87f79043d9c853c12)