Patch sos plugin for rhev
Until there is a community sos plugin for kimchi, this gathers
kimichi logs and config information.

The original internal plugin had a section to capture virsh
pool info, but that is duplicating the libvirt logs and xml files
gathered by the libvirt plugin. Removing this also takes care
of the issue that when running on RHEV, virsh commands should
use the --read-only flag.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>