Change systemd killmode to avoid kill Yum process
Kimchi has a problem when updating itself. Kimchid is restarted and all
children are kill, including Yum. When that happens, the updates are not
finished properly and the system ends up in an inconsistente state.
This patch changes the Systemd KillMode to "process", so, child will
remains in the system, letting yum finishes all transactions in
background, while kimchi restarts correctly.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>