Remove slash "/" filter in template name when create VM
Users are allowed to create or update template names with "slash",
this generates an error when he tries to create a guest with that
template because the guest paramters validation prohibits slashes
in templates names. This patch fixes this problem.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>
(cherry picked from commit 30cf31969286055b567c8a7ae54f0e2b4f214642)