Hot add/remove CPUs on PPC
* Use locks to prevent concurrent updates to VMs:
If several users try to update the same VM at the same time, an error
may happen due to concurrent access.
Define a dictionary containing one lock per each VM, and use it when
updating the VM: the lock is acquired before the update operation begins
and it is released when the operation finishes (whether by success or
failure).

* Update VCPU by using libvirt function:
Currently, the VCPU count is updated by editing the configuration XML
(tag: <vcpu>). However, in some cases, editing that tag will only update
the maximum VCPU count, not the current one. For example, if the VM has
the following XML tag:   <vcpu current='2'>8</vcpu>
and the user updates the VCPU value to 10, Kimchi will update the XML
tag to: <vcpu current='2'>10</vcpu>
which is probably not what the user wanted.
Use the libvirt function "setVcpusFlags" to update the current and the
maximum VCPU values to the one specified by the user.

* Set max VCPU count to the max supported value:
When updating the CPU count of a VM, its maximum CPU count is also
updated to the same value. This wouldn't allow for CPU hotplug
afterwards. Update the maximum CPU count to the maximum value
supported by libvirt when the user updates the VM CPU count.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>
Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>