Fix CPU Hotplug backend and topology settings
* This patch fixes several issues found in CPU Hotplug backend
side:
   - enables the backend feature;
   - fixes cpu offline update;
   - fixes vcpusFlags calls;
   - fixes error messages KCHVM0055E and KCHVM0054E;
   - fixes problems hotpluging more then 10 sockets;
   - fixes vcpus/maxvcpu management;

* This patch also changes vcpu setting when guest is being
created. Now 'current' is always set and is the total of cpus
from the Template and maxVcpu is the >= 255 or (sockets * cores
* threads) in PPC. In x86 it is the value of libvirt getMaxVcpu.

* This patch enables CPU hotplug when topology is set, following
this rules for PPC. If guest is running, new cpu value must be
a multiple of cores*threads configured, otherwise, a error is
returned. If guest if offline, then user is not allow to do any
change.

* Change NUMA management. Sets all cpus to 0. There is not
impact to the guest.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>