Change max memory limit to 4x original guest memory size
This patch changes the total maximum memory limit. Instead of total
hostt physical memory size, it sets a value equals 4 times the guest
memory configured in the template, when the guest is created.
This value does not change during memory update. It will only change
during update, if the host does not have the enough maxMemory size.
This new approach is necessary to avoid problems in live migrations and
if host physical memory decreases.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>