Limit number of log files kept in the system to 10
Currently Wok is configured to keep up to 1000 log files of 10MB,
this behavior makes Wok/Kimchi log space become huge (up to 20GB,
for error and access logs).
This patch changes the configuration, limiting the number of stored
log files to 10. So space required in the system will be up to 200MB.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>